+++ CHANGELOG.md
@@ -27,6 +27,7 @@
 * Fix Docker discovery of exposed ports on stopped containers.
 * Fix `aerospike upgrade` command to handle differing folder names in minor versions.
 * Fix `cluster partition list` race condition when in parallel mode.
+* Improve AWS expiry system refresh and GCP expiry system discovery.
 
 #### 7.1.1
 * GCP just made `DiscardLocalSsd` non-optional when stopping instances. Adjusting accordingly.